update constants building pattern
