+++ deprecated/parser.js
@@ -9,7 +9,7 @@ var spawn = cp.spawn;
 var exec = cp.exec;
 var bodyParser = require('body-parser');
 var progress = require('request-progress');
-//var constants = require('./constants.json');
+//var constants = require('./constants.js');
 var app = express();
 //var capacity = require('os').cpus().length;
 var capacity = Number(config.PARSER_PARALLELISM);
