+++ fullhistory.js
@@ -9,7 +9,7 @@ var db = require('./db');
 var getData = utility.getData;
 var queries = require('./queries');
 var insertMatch = queries.insertMatch;
-var constants = require('./constants.json');
+var constants = require('./constants.js');
 var urllib = require('url');
 var generateJob = utility.generateJob;
 var config = require('./config');
