+++ json/kill_streaks.json
@@ -1,0 +1,10 @@
+{
+  "3": "Killing Spree",
+  "4": "Dominating",
+  "5": "Mega Kill",
+  "6": "Unstoppable",
+  "7": "Wicked Sick",
+  "8": "Monster Kill",
+  "9": "Godlike",
+  "10": "Beyond Godlike"
+}
