+++ json/lane_role.json
@@ -1,0 +1,6 @@
+{
+  "1": "Safe",
+  "2": "Mid",
+  "3": "Off",
+  "4": "Jungle"
+}
