+++ json/leaver_status.json
@@ -1,0 +1,9 @@
+{
+  "0": "Played",
+  "1": "Left Safely",
+  "2": "Abandoned (DC)",
+  "3": "Abandoned",
+  "4": "Abandoned (AFK)",
+  "5": "Never Connected",
+  "6": "Never Connected (Timeout)"
+}
