+++ json/multi_kills.json
@@ -1,0 +1,6 @@
+{
+  "2": "Double Kill",
+  "3": "Triple Kill",
+  "4": "Ultra Kill",
+  "5": "Rampage"
+}
