+++ json/objectives.json
@@ -1,0 +1,9 @@
+{
+  "CHAT_MESSAGE_TOWER_KILL": "Tower",
+  "CHAT_MESSAGE_BARRACKS_KILL": "Barracks",
+  "CHAT_MESSAGE_ROSHAN_KILL": "Roshan",
+  "CHAT_MESSAGE_AEGIS": "Picked up the Aegis",
+  "CHAT_MESSAGE_FIRSTBLOOD": "First Blood",
+  "CHAT_MESSAGE_TOWER_DENY": "Tower Deny",
+  "CHAT_MESSAGE_AEGIS_STOLEN": "Stole the Aegis"
+}
