+++ json/order_types.json
@@ -1,0 +1,31 @@
+{
+  "0": "DOTA_UNIT_ORDER_NONE",
+  "1": "DOTA_UNIT_ORDER_MOVE_TO_POSITION",
+  "2": "DOTA_UNIT_ORDER_MOVE_TO_TARGET",
+  "3": "DOTA_UNIT_ORDER_ATTACK_MOVE",
+  "4": "DOTA_UNIT_ORDER_ATTACK_TARGET",
+  "5": "DOTA_UNIT_ORDER_CAST_POSITION",
+  "6": "DOTA_UNIT_ORDER_CAST_TARGET",
+  "7": "DOTA_UNIT_ORDER_CAST_TARGET_TREE",
+  "8": "DOTA_UNIT_ORDER_CAST_NO_TARGET",
+  "9": "DOTA_UNIT_ORDER_CAST_TOGGLE",
+  "10": "DOTA_UNIT_ORDER_HOLD_POSITION",
+  "11": "DOTA_UNIT_ORDER_TRAIN_ABILITY",
+  "12": "DOTA_UNIT_ORDER_DROP_ITEM",
+  "13": "DOTA_UNIT_ORDER_GIVE_ITEM",
+  "14": "DOTA_UNIT_ORDER_PICKUP_ITEM",
+  "15": "DOTA_UNIT_ORDER_PICKUP_RUNE",
+  "16": "DOTA_UNIT_ORDER_PURCHASE_ITEM",
+  "17": "DOTA_UNIT_ORDER_SELL_ITEM",
+  "18": "DOTA_UNIT_ORDER_DISASSEMBLE_ITEM",
+  "19": "DOTA_UNIT_ORDER_MOVE_ITEM",
+  "20": "DOTA_UNIT_ORDER_CAST_TOGGLE_AUTO",
+  "21": "DOTA_UNIT_ORDER_STOP",
+  "22": "DOTA_UNIT_ORDER_TAUNT",
+  "23": "DOTA_UNIT_ORDER_BUYBACK",
+  "24": "DOTA_UNIT_ORDER_GLYPH",
+  "25": "DOTA_UNIT_ORDER_EJECT_ITEM_FROM_STASH",
+  "26": "DOTA_UNIT_ORDER_CAST_RUNE",
+  "27": "DOTA_UNIT_ORDER_PING_ABILITY",
+  "28": "DOTA_UNIT_ORDER_MOVE_TO_DIRECTION"
+}
