+++ json/parse_status.json
@@ -1,0 +1,6 @@
+{
+  "0": "Queued",
+  "1": "Unavailable",
+  "2": "Parsed",
+  "3": "Skipped"
+}
