+++ json/patch.json
@@ -1,0 +1,202 @@
+[
+  {
+    "name": "6.85",
+    "date": "2015-09-24T20:00:00Z"
+  },
+  {
+    "name": "6.84c",
+    "date": "2015-05-18T16:00:00Z"
+  },
+  {
+    "name": "6.84b",
+    "date": "2015-05-09T18:00:00Z"
+  },
+  {
+    "name": "6.84",
+    "date": "2015-04-30T21:00:00Z"
+  },
+  {
+    "name": "6.83c",
+    "date": "2015-02-13T00:00:00Z"
+  },
+  {
+    "name": "6.83b",
+    "date": "2015-01-13T00:00:00Z"
+  },
+  {
+    "name": "6.83",
+    "date": "2014-12-17T00:00:00Z"
+  },
+  {
+    "name": "6.82c",
+    "date": "2014-10-15T00:00:00Z"
+  },
+  {
+    "name": "6.82b",
+    "date": "2014-09-28T00:00:00Z"
+  },
+  {
+    "name": "6.82",
+    "date": "2014-09-24T00:00:00Z"
+  },
+  {
+    "name": "6.81d",
+    "date": "2014-08-09T00:00:00Z"
+  },
+  {
+    "name": "6.81c",
+    "date": "2014-08-03T00:00:00Z"
+  },
+  {
+    "name": "6.81b",
+    "date": "2014-06-02T00:00:00Z"
+  },
+  {
+    "name": "6.81",
+    "date": "2014-04-29T00:00:00Z"
+  },
+  {
+    "name": "6.80c",
+    "date": "2014-03-03T00:00:00Z"
+  },
+  {
+    "name": "6.80b",
+    "date": "2014-03-02T00:00:00Z"
+  },
+  {
+    "name": "6.80",
+    "date": "2014-01-27T00:00:00Z"
+  },
+  {
+    "name": "6.79e",
+    "date": "2014-01-12T00:00:00Z"
+  },
+  {
+    "name": "6.79d",
+    "date": "2014-01-07T00:00:00Z"
+  },
+  {
+    "name": "6.79c",
+    "date": "2013-12-11T00:00:00Z"
+  },
+  {
+    "name": "6.79b",
+    "date": "2013-11-27T00:00:00Z"
+  },
+  {
+    "name": "6.79",
+    "date": "2013-11-24T00:00:00Z"
+  },
+  {
+    "name": "6.78c",
+    "date": "2013-06-30T00:00:00Z"
+  },
+  {
+    "name": "6.78b",
+    "date": "2013-06-03T00:00:00Z"
+  },
+  {
+    "name": "6.78",
+    "date": "2013-05-30T00:00:00Z"
+  },
+  {
+    "name": "6.77c",
+    "date": "2013-03-16T00:00:00Z"
+  },
+  {
+    "name": "6.77b",
+    "date": "2013-01-15T00:00:00Z"
+  },
+  {
+    "name": "6.77",
+    "date": "2012-12-15T00:00:00Z"
+  },
+  {
+    "name": "6.76c",
+    "date": "2012-10-28T00:00:00Z"
+  },
+  {
+    "name": "6.76b",
+    "date": "2012-10-22T00:00:00Z"
+  },
+  {
+    "name": "6.76",
+    "date": "2012-10-21T00:00:00Z"
+  },
+  {
+    "name": "6.75b",
+    "date": "2012-10-02T00:00:00Z"
+  },
+  {
+    "name": "6.75",
+    "date": "2012-09-30T00:00:00Z"
+  },
+  {
+    "name": "6.74c",
+    "date": "2012-03-21T00:00:00Z"
+  },
+  {
+    "name": "6.74b",
+    "date": "2012-03-20T00:00:00Z"
+  },
+  {
+    "name": "6.74",
+    "date": "2012-03-10T00:00:00Z"
+  },
+  {
+    "name": "6.73c",
+    "date": "2012-01-10T00:00:00Z"
+  },
+  {
+    "name": "6.73b",
+    "date": "2011-12-28T00:00:00Z"
+  },
+  {
+    "name": "6.73",
+    "date": "2011-12-24T00:00:00Z"
+  },
+  {
+    "name": "6.72f",
+    "date": "2011-07-22T00:00:00Z"
+  },
+  {
+    "name": "6.72e",
+    "date": "2011-07-18T00:00:00Z"
+  },
+  {
+    "name": "6.72d",
+    "date": "2011-07-17T00:00:00Z"
+  },
+  {
+    "name": "6.72c",
+    "date": "2011-06-03T00:00:00Z"
+  },
+  {
+    "name": "6.72b",
+    "date": "2011-05-04T00:00:00Z"
+  },
+  {
+    "name": "6.72",
+    "date": "2011-04-27T00:00:00Z"
+  },
+  {
+    "name": "6.71b",
+    "date": "2011-01-23T00:00:00Z"
+  },
+  {
+    "name": "6.71",
+    "date": "2011-01-21T00:00:00Z"
+  },
+  {
+    "name": "6.70c",
+    "date": "2010-12-30T00:00:00Z"
+  },
+  {
+    "name": "6.70b",
+    "date": "2010-12-28T00:00:00Z"
+  },
+  {
+    "name": "6.70",
+    "date": "2010-12-24T00:00:00Z"
+  }
+]
