+++ json/player_colors.json
@@ -1,0 +1,12 @@
+{
+  "0": "#2E6AE6",
+  "1": "#5DE6AD",
+  "2": "#AD00AD",
+  "3": "#DCD90A",
+  "4": "#E66200",
+  "128": "#E67AB0",
+  "129": "#92A440",
+  "130": "#5CC5E0",
+  "131": "#00771F",
+  "132": "#956000"
+}
