+++ json/runes.json
@@ -1,0 +1,26 @@
+{
+  "0": {
+    "name": "DD",
+    "img": ""
+  },
+  "1": {
+    "name": "HST",
+    "img": ""
+  },
+  "2": {
+    "name": "ILL",
+    "img": ""
+  },
+  "3": {
+    "name": "INV",
+    "img": ""
+  },
+  "4": {
+    "name": "REG",
+    "img": ""
+  },
+  "5": {
+    "name": "BTY",
+    "img": ""
+  }
+}
