+++ json/skill.json
@@ -1,0 +1,5 @@
+{
+  "1": "Normal",
+  "2": "High",
+  "3": "Very High"
+}
