+++ json/skillshots.json
@@ -1,0 +1,6 @@
+{
+  "mirana_arrow": true,
+  "pudge_meat_hook": true,
+  "rattletrap_hookshot": true,
+  "invoker_sun_strike": true
+}
