+++ json/xp_level.json
@@ -1,0 +1,27 @@
+[
+  0,
+  200,
+  400,
+  900,
+  1400,
+  2000,
+  2600,
+  3200,
+  4400,
+  5400,
+  6000,
+  8200,
+  9000,
+  10400,
+  11900,
+  13500,
+  15200,
+  17000,
+  18900,
+  20900,
+  23000,
+  25200,
+  27500,
+  29900,
+  32400
+]
