+++ package.json
@@ -9,13 +9,13 @@
     "dev2": "pm2-dev deploy.json",
     "test": "rm -rf testfiles && git clone https://github.com/yasp-dota/testfiles && npm run build && istanbul cover _mocha --report lcovonly -- -R spec",
     "coveralls": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage",
+    "constants": "node constantsUpdate.js",
     "update": "npm-check-updates -u",
-    "constants": "node runner.js constants",
     "webpack": "webpack -p",
     "watch": "webpack --watch",
     "mocha": "mocha",
     "maven": "mvn -q -f java_parser/pom.xml clean install -U",
-    "build": "npm install && npm run maven && npm run webpack && npm run constants",
+    "build": "npm install && npm run maven && npm run webpack",
     "startup": "sudo pm2 startup ubuntu -u yasp",
     "create": "sudo -u postgres psql -f sql/create.sql",
     "deploy-core": "node deploy core",
