+++ routes/players.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var players = express.Router();
 var async = require('async');
-var constants = require('../constants.json');
+var constants = require('../constants.js');
 var queries = require("../queries");
 var utility = require('../utility');
 var generatePositionData = utility.generatePositionData;
